Refactor: 불필요한 코드 제거

+++ orders/views.py
@@ -44,23 +44,17 @@ class BasketView(View):
 
     @login_decorator
     def delete(self, request, order_item):
-        try:
-            order_item = OrderItem.objects.get(id=order_item)
+        if not OrderItem.objects.filter(id=order_item).exists:
+            return JsonResponse({'message':'INVALID_ORDER_ITEMS'}, status=400)
+            
+        order_item = OrderItem.objects.get(id=order_item)
 
-            if order_item.order.user != request.user:
-                return JsonResponse({'message':'INVALID_ORDER_ITEMS'}, status=401)
+        if order_item.order.user != request.user:
+            return JsonResponse({'message':'INVALID_USER'}, status=401)
 
-            if order_item.order.order_status.id != OrderStatus.BASKET:
-                return JsonResponse({'message':'ORDER_DOSE_NOT_EXIST'}, status=400)
+        if order_item.order.order_status.id != OrderStatus.BASKET:
+            return JsonResponse({'message':'ORDER_DOSE_NOT_EXIST'}, status=400)
 
-            order_item.delete()
+        order_item.delete()
 
-            return JsonResponse({'message':'SUCCESS'}, status=204)
-
-        except MultipleObjectsReturned:
-            return JsonResponse({'message':'MULTIPLE_OBJECT_RETURNED'}, status=400)
-
-        except ObjectDoesNotExist:
-            return JsonResponse({'message':'DOES_NOT_EXIST'}, status=400)
-
-        
+        return JsonResponse({'message':'SUCCESS'}, status=204)
